girrepository: Replace usage of gsize with standard size_t
We just do a safe s/gsize/size_t/ replacement here without doing any
changes to places in which different size of size_t and gsize may be
actually different and create troubles.

+++ girepository/girmodule.c
@@ -319,8 +319,8 @@ gi_ir_module_build_typelib (GIIrModule *module)
 {
   GError *error = NULL;
   GITypelib *typelib;
-  gsize length;
-  unsigned int i;
+  size_t length;
+  size_t i;
   GList *e;
   Header *header;
   DirEntry *entry;
